core: wifi: Prevent rescan without need
Scanning networks is a operation that thakes some time.
With currently logic, if there's more than one service asking for the available networks, more than one scan operation will be performed, even if it was already scanning and the result is already available.
By introducing a `scanning` state, we answer fetchs with the same result for everybody that asks for the available networks if a scan operation was already in place.