core: ardupilot-manager: Allow to plug more than one serial board
This commit changes the way serial devices are detected.

Instead of relying on UDEV rules it uses `pyserial` to detect available serial devices.

With UDEV rules, we can only detect one at a given time. With this changes, this limitation no longer occures.

The validation rules (manufacturer and product name) are kept.